feat: full crud implementation on the controller

+++ src/main/java/com/devraul/crud_demo/dao/EmployeeDAO.java
@@ -6,7 +6,7 @@ import java.util.List;
 
 public interface EmployeeDAO {
     List<Employee> findAll();
-    Employee findById(int id);
+    Employee findById(Long id);
     Employee save(Employee employee);
-    void deleteById(int id);
+    void deleteById(Long id);
 }
